pyproject.toml
[project]
name = "django-api-online-shopping"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "MELJOESTEIN", email = "thierrymemel@gmail.com" }
]
dependencies = [
    "django>=5.1.2",
    "djangorestframework>=3.15.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[tool.rye]
managed = true
virtual = true
dev-dependencies = []
